backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1

# Database and ORM
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9  # Commented out for Windows compatibility
alembic==1.12.1
asyncpg==0.29.0

# Data Validation and Settings
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=42.0.0
bcrypt==4.1.2

# Environment and Configuration
python-dotenv==1.0.0

# Background Tasks and Scheduling
redis==5.0.1
celery==5.3.4
APScheduler==3.10.4
schedule==1.2.0

# SNMP and Network Management
pysnmp==4.4.12
pysnmp-mibs==0.1.6

# Data Analysis and Visualization
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# HTTP Client and Utilities
httpx==0.25.2

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Document Generation and Templates
Jinja2==3.1.2
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7

# Image Processing and QR Codes
qrcode[pil]==7.4.2
Pillow==10.1.0

# Additional Dependencies for Production
gunicorn==21.2.0
starlette==0.27.0
typing-extensions==4.8.0
python-dateutil==2.8.2